+++ modules/enumerations.js
@@ -1,0 +1,99 @@
+const enumerations = (function () {
+
+  const soilLevel = {
+    0: 'extra light',
+    1: 'light',
+    2: 'normal',
+    3: 'heavy',
+    4: 'extra heavy'
+  }
+
+  const spinLevel = {
+    0: 'no spin',
+    1: 'unused',
+    2: 'medium',
+    3: 'high',
+    4: 'extra high',
+    5: 'disabled'
+  }
+
+  const waterTemp = {
+    21: 'hot',
+    20: 'warm',
+    19: 'colors',
+    18: 'cool',
+    17: 'cold',
+    16: 'tap cold'
+  }
+
+  const cycle = {
+    0: 'blank',
+    1: 'basket_clean',
+    2: 'rinse_and_spin',
+    3: 'quick_rinse',
+    4: 'bulky_items',
+    5: 'sanitize',
+    6: 'towels_and_sheets',
+    7: 'steam_refresh',
+    8: 'normal',
+    9: 'whites',
+    10: 'darks',
+    11: 'jeans',
+    12: 'hand_wash',
+    13: 'delicates',
+    14: 'speed_wash',
+    15: 'heavy_duty',
+    16: 'allergen',
+    17: 'power_clean',
+    18: 'rinse_and_spin',
+    19: 'single_item',
+    20: 'colors',
+    21: 'cold_wash',
+    128: 'cottons',
+    129: 'easy_care',
+    130: 'active_wear',
+    131: 'time_dry',
+    132: 'dewrinkle',
+    133: 'air_fluff',
+    134: 'steam_refresh',
+    135: 'steam_dewrinkle',
+    136: 'speed_dry',
+    137: 'mixed',
+    138: 'speed_dry',
+    139: 'casuals',
+    140: 'warm_up',
+    141: 'energy_saver'
+  }
+
+  const dryTemp = {
+    1: 'no heat',
+    2: 'low',
+    3: 'medium',
+    4: 'high'
+  }
+
+  const stainPretreat = {
+    0: 'off',
+    1: 'tomato',
+    2: 'wine',
+    3: 'blood',
+    4: 'grass',
+    5: 'dirt'
+  }
+
+  function makeReadable (text) {
+    return text.replace(/[_-]/g, ' ')
+  }
+
+  return {
+    spinLevel,
+    soilLevel,
+    waterTemp,
+    cycle,
+    dryTemp,
+    stainPretreat,
+    makeReadable
+  }
+})()
+
+module.exports = enumerations
