+++ modules/talkback.js
@@ -3,6 +3,7 @@ const adapter = require('gea-adapter-usb')
 const say = require('say')
 const eventbuffer = require('./eventbuffer')
 const erds = require('./erds')
+const enums = require('./enumerations')
 
 const talkback = (function () {
   const app = gea.configure({
@@ -137,24 +138,24 @@ const talkback = (function () {
   function handleCycleSelected (event, appliance) {
     let newCycle = erds.erd(erds.CYCLE_SELECTED).data(event)
     if (newCycle !== appliance.oldCycle) {
-      say.speak('cycle: '+newCycle, 'voice_us2_mbrola')
+      say.speak(enums.makeReadable(enums.cycle[newCycle]), 'voice_us2_mbrola')
     }
     appliance.oldCycle = newCycle
   }
 
   function handleWaterTemp (event, appliance) {
     let waterTemp = erds.erd(erds.WATER_TEMP).data(event)
-    say.speak('water temp: '+waterTemp, 'voice_us2_mbrola')
+    say.speak('water temp: '+enums.waterTemp[waterTemp], 'voice_us2_mbrola')
   }
 
   function handleSpinLevel (event, appliance) {
     let spinLevel = erds.erd(erds.SPIN_LEVEL).data(event)
-    say.speak('spin level: '+spinLevel, 'voice_us2_mbrola')
+    say.speak('spin level: '+enums.spinLevel[spinLevel], 'voice_us2_mbrola')
   }
 
   function handleSoilLevel (event, appliance) {
     let soilLevel = erds.erd(erds.SPIN_LEVEL).data(event)
-    say.speak('soil level: '+soilLevel, 'voice_us2_mbrola')
+    say.speak('soil level: '+enums.soilLevel[soilLevel], 'voice_us2_mbrola')
   }
 
   function handleMachineStatus (event, appliance) {
@@ -179,12 +180,12 @@ const talkback = (function () {
 
   function handleDryTemp (event, appliances) {
     let temp = erds.erd(erds.DRY_TEMP).data(event)
-    say.speak('dry temp: '+temp, 'voice_us2_mbrola')
+    say.speak('dry temp: '+enums.dryTemp[temp], 'voice_us2_mbrola')
   }
 
   function handleStainPretreat (event, appliance) {
     let level = erds.erd(erds.STAIN_PRETREAT).data(event)
-    say.speak('stain pretreat: '+level, 'voice_us2_mbrola')
+    say.speak('stain pretreat: '+enums.stainPretreat[level], 'voice_us2_mbrola')
   }
 
   return {
